chore: move header inside main-layout folder
