+++ frontend/src/app/main-layout/header/header.component.scss
