+++ frontend/src/app/main-layout/main-layout.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 
-import { HeaderComponent } from '../components/header';
+import { HeaderComponent } from './header';
 import { SidebarComponent } from './sidebar';
 
 @Component({
